src/requirements.txt
packaging==25.0
PyYAML==6.0.2
Requests==2.32.4
setuptools==80.9.0
